# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Qema Al Nathair Travel & Tourism API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/signup:
    post:
      operationId: signup
      tags: [auth]
      summary: Create a customer account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SignupInput"
      responses:
        "201":
          description: Account created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Log in with phone or email
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /auth/me:
    get:
      operationId: getCurrentUser
      tags: [auth]
      summary: Get current logged-in user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/profile:
    patch:
      operationId: updateProfile
      tags: [auth]
      summary: Update current user's profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProfileUpdate"
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /auth/forgot-password:
    post:
      operationId: requestPasswordReset
      tags: [auth]
      summary: Request a password reset
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PasswordResetRequest"
      responses:
        "200":
          description: Reset instructions issued
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /auth/reset-password:
    post:
      operationId: resetPassword
      tags: [auth]
      summary: Reset password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PasswordReset"
      responses:
        "200":
          description: Password reset
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"

  # ---------- Home ----------
...
  /admin/banners:
    get:
      operationId: listAllBanners
      tags: [admin]
      summary: List all banners (admin)
      responses:
        "200":
          description: Banners
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Banner"
    post:
      operationId: createBanner
      tags: [admin]
      summary: Create a banner (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BannerInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Banner"
  /admin/banners/{id}:
    patch:
      operationId: updateBanner
      tags: [admin]
      summary: Update a banner (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BannerUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Banner"
    delete:
      operationId: deleteBanner
      tags: [admin]
      summary: Delete a banner (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ---------- Testimonials ----------
...
  /admin/customers/{id}:
    get:
      operationId: getCustomer
      tags: [admin]
      summary: Get a customer (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Customer
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required:
        - message
    UserRole:
      type: string
      enum: [customer, admin]
    User:
      type: object
      required: [id, fullName, phone, role, language, currency, createdAt]
      properties:
        id: { type: integer }
        fullName: { type: string }
        phone: { type: string }
        email: { type: ["string", "null"] }
        avatarUrl: { type: ["string", "null"] }
        role:
          $ref: "#/components/schemas/UserRole"
        language: { type: string }
        currency: { type: string }
        createdAt: { type: string, format: date-time }
    SignupInput:
      type: object
      required: [fullName, phone, password]
      properties:
        fullName: { type: string, minLength: 1 }
        phone: { type: string, minLength: 3 }
        email: { type: string }
        password: { type: string, minLength: 6 }
    LoginInput:
      type: object
      required: [identifier, password]
      properties:
        identifier:
          type: string
          description: Phone number or email
        password: { type: string }
    ProfileUpdate:
      type: object
      properties:
        fullName: { type: string, minLength: 1 }
        email: { type: string }
        phone: { type: string }
        avatarUrl: { type: string }
        language: { type: string }
        currency: { type: string }
    PasswordResetRequest:
      type: object
      required: [identifier]
      properties:
        identifier: { type: string }
    PasswordReset:
      type: object
      required: [identifier, newPassword]
      properties:
        identifier: { type: string }
        newPassword: { type: string, minLength: 6 }
...
    Banner:
      type: object
      required: [id, imageUrl, sortOrder, isActive, createdAt]
      properties:
        id: { type: integer }
        imageUrl: { type: string }
        title: { type: ["string", "null"] }
        linkUrl: { type: ["string", "null"] }
        sortOrder: { type: integer }
        isActive: { type: boolean }
        createdAt: { type: string, format: date-time }
    BannerInput:
      type: object
      required: [imageUrl]
      properties:
        imageUrl: { type: string }
        title: { type: string }
        linkUrl: { type: string }
        sortOrder: { type: integer }
        isActive: { type: boolean }
    BannerUpdate:
      type: object
      properties:
        imageUrl: { type: string }
        title: { type: string }
        linkUrl: { type: string }
        sortOrder: { type: integer }
        isActive: { type: boolean }
...