# lib/api-zod/src/generated/types/userRole.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * Qema Al Nathair Travel & Tourism API
 * OpenAPI spec version: 0.1.0
 */

export type UserRole = typeof UserRole[keyof typeof UserRole];


export const UserRole = {
  customer: 'customer',
  admin: 'admin',
} as const;